+++ client/src/features/timeline/List.tsx
@@ -12,6 +12,7 @@ import {
   ITimelinesTransformed,
 } from '../../types/ITimelines'
 import { IHalLinks } from '../../types/IHalLinks'
+import OverflowScrollContainer from '../common/OverflowScrollContainer'
 
 import ListRow from './ListRow'
 
@@ -68,8 +69,11 @@ const List: React.FC<IProps> = ({
 
   const showEras = TimelineParser.showYearEra(yearsArray)
   return (
-    <React.Fragment>
-      <dl data-testid="timeline-list-container">
+    <OverflowScrollContainer
+      style={{ height: '100%' }}
+      className="timelineListOverflowScrollContainer"
+    >
+      <dl id="timeline-list-container" data-testid="timeline-list-container">
         {sortedYearsRange.slice(0, displayLength).map((year) => (
           <StyledDiv key={year} className="mb-2">
             <HoverableRow>
@@ -125,7 +129,7 @@ const List: React.FC<IProps> = ({
           Show Less
         </button>
       )}
-    </React.Fragment>
+    </OverflowScrollContainer>
   )
 }